Feat - Multi-year support for time-based entries
This commit adds year to time entries, which can be entered on different years, such as month and week.
The commit solves issue with having multiple entries for same week or month number. In order to do that, the commit
updates the week type which brakes back-wards campatibility in serialization.