Feat - Semi automatic backups
This commit adds the functionality to enable semi-automatic backups of the
stored data. Reason for it to be semi-automated is that it will do it without
the user, based on a time interval. However, the butler still need to be executed since
it's not running as a daemon. Time data is stored in a deticated backup directory.
The interval and location can be changed from the config file.

A force trigger is also added which let's the user trigger a backup from the cli directly as a
command.